代码调试

+++ qk-dam/qike-dm-datasource/src/main/java/com/qk/dm/datasource/service/impl/DsDirServiceImpl.java
@@ -1,16 +1,26 @@
 package com.qk.dm.datasource.service.impl;
 
+import com.google.gson.reflect.TypeToken;
 import com.qk.dam.commons.exception.BizException;
+import com.qk.dam.commons.util.GsonUtil;
+import com.qk.dam.datasource.entity.*;
+import com.qk.dam.datasource.enums.ConnTypeEnum;
+import com.qk.dm.datasource.constant.DsConstant;
+import com.qk.dm.datasource.entity.DsDatasource;
 import com.qk.dm.datasource.entity.DsDir;
 import com.qk.dm.datasource.entity.QDsDir;
+import com.qk.dm.datasource.mapstruct.mapper.DSDatasourceMapper;
 import com.qk.dm.datasource.mapstruct.mapper.DsDirMapper;
+import com.qk.dm.datasource.repositories.DsDatasourceRepository;
 import com.qk.dm.datasource.repositories.DsDirRepository;
 import com.qk.dm.datasource.service.DsDirService;
+import com.qk.dm.datasource.vo.DsDatasourceVO;
 import com.qk.dm.datasource.vo.DsDirReturnVO;
 import com.qk.dm.datasource.vo.DsDirVO;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
+import org.springframework.util.StringUtils;
 
 import java.util.*;
 
@@ -25,9 +35,11 @@ import java.util.*;
 public class DsDirServiceImpl implements DsDirService {
   private final QDsDir qDsDir = QDsDir.dsDir;
   private final DsDirRepository dsDirRepository;
+  private final DsDatasourceRepository dsDatasourceRepository;
 
-  public DsDirServiceImpl(DsDirRepository dsDirRepository) {
+  public DsDirServiceImpl(DsDirRepository dsDirRepository, DsDatasourceRepository dsDatasourceRepository) {
     this.dsDirRepository = dsDirRepository;
+    this.dsDatasourceRepository = dsDatasourceRepository;
   }
 
   @Override
@@ -57,9 +69,21 @@ public class DsDirServiceImpl implements DsDirService {
     }
     ids.add(id);
     getIds(ids, id);
-    // 批量删除
-    Iterable<DsDir> dsDirList = dsDirRepository.findAll(qDsDir.id.in(ids));
-    dsDirRepository.deleteAll(dsDirList);
+    //查询目录下是否存在数据源信息
+    List<DsDatasourceVO> dsDatasourceVOList = new ArrayList<>();
+    ids.forEach(
+            dirid ->{
+              List<DsDatasourceVO> list = getDataSourceList(id);
+              dsDatasourceVOList.addAll(list);
+            }
+    );
+    if (!CollectionUtils.isEmpty(dsDatasourceVOList)){
+      throw new BizException("存在数据源连接信息，请先删除！！！");
+    }else {
+      // 批量删除
+      Iterable<DsDir> dsDirList = dsDirRepository.findAll(qDsDir.id.in(ids));
+      dsDirRepository.deleteAll(dsDirList);
+    }
   }
 
   @Override
@@ -87,23 +111,83 @@ public class DsDirServiceImpl implements DsDirService {
     return trees;
   }
 
+  private DsDirReturnVO findChildren(DsDirReturnVO dsDirReturnVO, List<DsDirReturnVO> dsDirVOList) {
+      dsDirReturnVO.setChildren(new ArrayList<>());
+      for (DsDirReturnVO DSDTV : dsDirVOList) {
+        if (dsDirReturnVO.getKey().equals(DSDTV.getParentId())) {
+          if (dsDirReturnVO.getChildren() == null) {
+            dsDirReturnVO.setChildren(new ArrayList<>());
+          }
+          dsDirReturnVO.getChildren().add(findChildren(DSDTV, dsDirVOList));
+        }
+      }
+      return dsDirReturnVO;
+  }
+
   /**
-   * 递归查找子类
+   * 递归查找子类并查询目录下对应的数据源信息
    * @param dsDirReturnVO
    * @param dsDirVOList
    * @return
    */
-  private DsDirReturnVO findChildren(DsDirReturnVO dsDirReturnVO, List<DsDirReturnVO> dsDirVOList) {
+  private DsDirReturnVO findDirChildren(DsDirReturnVO dsDirReturnVO, List<DsDirReturnVO> dsDirVOList) {
+    dsDirReturnVO.setTyep(DsConstant.DIR_TYPE);
     dsDirReturnVO.setChildren(new ArrayList<>());
     for (DsDirReturnVO DSDTV : dsDirVOList) {
       if (dsDirReturnVO.getKey().equals(DSDTV.getParentId())) {
         if (dsDirReturnVO.getChildren() == null) {
           dsDirReturnVO.setChildren(new ArrayList<>());
         }
+        List<DsDatasourceVO> datasourceList =getDataSourceList(dsDirReturnVO.getKey());
+        dsDirReturnVO.setDatasourceVOList(datasourceList);
         dsDirReturnVO.getChildren().add(findChildren(DSDTV, dsDirVOList));
       }
     }
     return dsDirReturnVO;
+  }
+
+  private List<DsDatasourceVO> getDataSourceList(Integer dicid) {
+    List<DsDatasourceVO> dsDatasourceVOList = new ArrayList<>();
+    if (!StringUtils.isEmpty(dicid)){
+      List<DsDatasource> byDicIdList = dsDatasourceRepository.getByDicId(Integer.toString(dicid));
+      if (byDicIdList != null) {
+        byDicIdList.forEach(
+                dsDatasource -> {
+                  DsDatasourceVO dsDatasourceVO = DSDatasourceMapper.INSTANCE.useDsDatasourceVO(dsDatasource);
+                  dsDatasourceVO.setTyep(DsConstant.DATASOURCE_TYPE);
+                  ConnectBasicInfo dsConnectBasicInfo = getConnectInfo(dsDatasource.getLinkType(), dsDatasource);
+                  dsDatasourceVO.setConnectBasicInfo(dsConnectBasicInfo);
+                  dsDatasourceVOList.add(dsDatasourceVO);
+                }
+        );
+      }
+    }
+    return dsDatasourceVOList;
+  }
+
+  private ConnectBasicInfo getConnectInfo(String type, DsDatasource dsDatasource) {
+    ConnectBasicInfo connectBasicInfo = null;
+    if (type.equalsIgnoreCase(ConnTypeEnum.MYSQL.getName())) {
+      String dataSourceValues = dsDatasource.getDataSourceValues();
+      connectBasicInfo =
+              GsonUtil.fromJsonString(dataSourceValues, new TypeToken<MysqlInfo>() {}.getType());
+    }
+    if (type.equalsIgnoreCase(ConnTypeEnum.HIVE.getName())) {
+      String dataSourceValues = dsDatasource.getDataSourceValues();
+      connectBasicInfo =
+              GsonUtil.fromJsonString(dataSourceValues, new TypeToken<HiveInfo>() {}.getType());
+    }
+    if (type.equalsIgnoreCase(ConnTypeEnum.ORACLE.getName())) {
+      String dataSourceValues = dsDatasource.getDataSourceValues();
+      connectBasicInfo =
+              GsonUtil.fromJsonString(dataSourceValues, new TypeToken<OracleInfo>() {}.getType());
+    }
+    if (type.equalsIgnoreCase(ConnTypeEnum.POSTGRESQL.getName())) {
+      String dataSourceValues = dsDatasource.getDataSourceValues();
+      connectBasicInfo =
+              GsonUtil.fromJsonString(dataSourceValues, new TypeToken<PostgresqlInfo>() {}.getType());
+    }
+    return connectBasicInfo;
   }
 
   /**
@@ -123,6 +207,53 @@ public class DsDirServiceImpl implements DsDirService {
             this.getDsdId(dsDicIdSet, Integer.toString(dsDir.getId()));
           });
     }
+  }
+
+  @Override
+  public List<DsDirReturnVO> getDsDirDataSource() {
+    List<DsDir> dsDirList = dsDirRepository.findAll();
+    List<DsDirReturnVO> dsDirVOList = new ArrayList<>();
+    if (!CollectionUtils.isEmpty(dsDirList)) {
+      dsDirList.forEach(
+              dsDir -> {
+                DsDirReturnVO dsDirReturnVO = DsDirMapper.INSTANCE.useDsDirVO(dsDir);
+                dsDirReturnVO.setKey(dsDir.getId());
+                dsDirReturnVO.setTitle(dsDir.getDicName());
+                dsDirVOList.add(dsDirReturnVO);
+              });
+    } else {
+      throw new BizException("获取目录为空");
+    }
+    return buildByRecursives(dsDirVOList);
+  }
+
+  /**
+   * 修改目录名称
+   * @param dsDirVO
+   */
+  @Override
+  public void updateDsDir(DsDirVO dsDirVO) {
+    DsDir dsDir = DsDirMapper.INSTANCE.useDsDir(dsDirVO);
+    dsDir.setGmtModified(new Date());
+    BooleanExpression predicate = qDsDir.id.eq(dsDir.getId());
+    boolean exists = dsDirRepository.exists(predicate);
+    if (exists){
+      dsDirRepository.saveAndFlush(dsDir);
+    }else{
+      throw new BizException(
+              "当前要更新的目录ID为："
+                      + dsDir.getId()
+                      + "目录名称为:"
+                      + dsDir.getDicName()
+                      + " 的数据，不存在！！！");
+    }
+  }
+
+  private List<DsDirReturnVO> buildByRecursives(List<DsDirReturnVO> dsDirVOList) {
+    DsDirReturnVO dsDirReturnVO=DsDirReturnVO.builder().key(0).title("全部数据源").build();
+    List<DsDirReturnVO> trees = new ArrayList<>();
+    trees.add(findDirChildren(dsDirReturnVO, dsDirVOList));
+    return trees;
   }
 
   /**
